# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Property Scan API
  version: "0.1"
  description: >-
    Multi-tenant construction-capture API. Canonical geometry is metric; imperial
    values are display-only. Measurements are preliminary estimates unless marked
    field_verified. Request/response field shapes are authoritatively defined by
    the zod schemas in packages/contracts; this document is the HTTP contract
    (paths, methods, auth, status codes) and is verified against the running
    router by a contract test.
servers:
  - url: "{apiBaseUrl}"
    variables:
      apiBaseUrl:
        default: http://localhost:4000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Clerk session JWT (production) or dev_<userId> token (development).
  parameters:
    organizationHeader:
      name: X-Organization-Id
      in: header
      required: true
      schema: { type: string, format: uuid }
      description: >-
        Selects among the caller's organizations; validated server-side against
        membership. Never authorization by itself.
...
  responses:
    Problem:
      description: RFC 9457 problem
      content:
        application/problem+json:
          schema:
            type: object
            required: [title, status]
            properties:
              type: { type: string }
              title: { type: string }
              status: { type: integer }
              detail: { type: string }
paths:
...
  /v1/media/uploads:
    post:
      summary: Register media upload (jpeg/png/heic); server-chosen object key
      parameters: [{ $ref: "#/components/parameters/organizationHeader" }]
      responses:
        "201": { description: Upload slot }
        "413": { description: Exceeds size limit }
  /v1/media/uploads/{mediaId}/content:
    put:
      summary: Media byte upload (local storage driver only)
      parameters: [{ $ref: "#/components/parameters/organizationHeader" }]
      responses: { "204": { description: Stored } }
  /v1/media/uploads/{mediaId}/complete:
    post:
      summary: >-
        Validate media — checksum, MIME signature vs declared type, dimensions;
        JPEG Exif APP1 removed before the object becomes ready.
      parameters: [{ $ref: "#/components/parameters/organizationHeader" }]
      responses:
        "200": { description: Ready }
        "409": { description: Bytes missing }
        "422": { description: Checksum or content-type mismatch (media rejected) }
  /v1/media/{mediaId}/content:
    get:
      summary: Authorized download — presigned redirect (S3) or direct stream (fs)
      parameters: [{ $ref: "#/components/parameters/organizationHeader" }]
      responses:
        "200": { description: Bytes }
        "302": { description: Redirect to short-lived signed URL }
        "404": { $ref: "#/components/responses/Problem" }
  /v1/openings/{openingId}/media-links:
    post:
      summary: Link a ready media item to an opening (ordered by position)
      parameters: [{ $ref: "#/components/parameters/organizationHeader" }]
      responses: { "201": { description: Linked } }
    get:
      summary: List opening photos with short-lived download URLs
      parameters: [{ $ref: "#/components/parameters/organizationHeader" }]
      responses: { "200": { description: OK } }